% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cade.R
\docType{data}
\name{cade_palette}
\alias{cade_palette}
\title{Cade Inspired Color Palette}
\format{
character vector of hex code strings
}
\usage{
cade_palette
}
\description{
Cade Inspired Color Palette
}
\examples{
cade_palette

}
\concept{cade}
\keyword{datasets}
