% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cade.R
\docType{data}
\name{cade_theme_colors}
\alias{cade_theme_colors}
\title{cade Inspired Theme Color Palette}
\format{
character vector of hex code strings
}
\usage{
cade_theme_colors
}
\description{
cade Inspired Theme Color Palette
}
\examples{
cade_theme_colors

}
\concept{cade}
\keyword{datasets}
