% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cade.R
\name{scale_color_cade_c}
\alias{scale_color_cade_c}
\alias{scale_fill_cade_c}
\alias{scale_color_cade_b}
\alias{scale_fill_cade_b}
\alias{scale_color_cade_d}
\alias{scale_fill_cade_d}
\alias{scale_colour_cade_d}
\alias{scale_colour_cade_c}
\alias{scale_colour_cade_b}
\title{cade Inspired Color Scales}
\usage{
scale_color_cade_c(...)

scale_fill_cade_c(...)

scale_color_cade_b(...)

scale_fill_cade_b(...)

scale_color_cade_d(...)

scale_fill_cade_d(...)

scale_colour_cade_d(...)

scale_colour_cade_c(...)

scale_colour_cade_b(...)
}
\arguments{
\item{...}{Additional arguments to pass to `ggplot2::binned_scale` for `_b`,
`ggplot2::scale_[fill/color]_gradient` for `_c`, or `ggplot2::discrete_scale`
for `_d`}
}
\value{
a `ggplot` scale object
}
\description{
cade Inspired Color Scales
}
\examples{
library(ggplot2)

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  facet_wrap(~(hwy > 29)) +
  theme_cade() +
  scale_fill_cade_d()

}
