% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cade.R
\name{theme_cade}
\alias{theme_cade}
\title{cade Inspired Theme}
\usage{
theme_cade(cade_font = FALSE, ...)
}
\arguments{
\item{cade_font}{should `theme_cade` use custom font? Default is `TRUE`.}

\item{...}{additional parameters to pass to `ggplot2::theme()`}
}
\value{
a `ggplot2` `theme` element
}
\description{
cade Inspired Theme
}
\examples{
library(ggplot2)

ggplot(data = data.frame(x = rnorm(50, 0, 1), y = rnorm(50,0,1)), aes(x = x, y = y)) +
  geom_smooth(method = 'lm') +
  geom_point() +
  labs(title = 'cade Scatter Plot') +
  theme_cade()

ggplot(mpg, aes(cty)) +
geom_density(aes(fill=factor(cyl)), alpha=0.8) +
  labs(title="Density plot",
       subtitle="City Mileage Grouped by Number of cylinders",
       caption="Source: mpg",
       x="City Mileage",
       fill="# Cylinders") +
  theme_cade() +
  scale_fill_cade_d()


}
\concept{cade}
