% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cade.R
\name{theme_cade}
\alias{theme_cade}
\title{Tema do Cade}
\usage{
theme_cade(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\description{
Clean ggplot theme with no panel background, black axis lines
  and grey fill colour for chart elements.
}
\author{
Tomas Barcellos \email{tomas.barcellos@cade.gov.br}
}
\concept{themes}
